Add bucket param to requred functions. Move functions to the file that they are used in.
